refactor(account): add index files
