+++ packages/account/activate/index.ts
@@ -1,0 +1,3 @@
+export * from './activate.component';
+export * from './activate.service';
+export * from './activate.route';
