+++ packages/account/help/index.ts
@@ -1,0 +1,2 @@
+export * from './healp.route';
+export * from './help.component';
