+++ packages/account/idp-callback/index.ts
@@ -1,0 +1,2 @@
+export * from './idp-callback.component';
+export * from './idp-callback.route';
