+++ packages/account/index.ts
@@ -1,31 +1,12 @@
-export { ActivateComponent } from './activate/activate.component';
-export { Activate } from './activate/activate.service';
-export { activateRoute } from './activate/activate.route';
-export { IHelpConfig, HelpComponent } from './help/help.component';
-export { helpRoute } from './help/healp.route';
-export { PasswordComponent } from './password/password.component';
-export { Password } from './password/password.service';
-export { PasswordModule } from './password/password.module';
-export { passwordRoute } from './password/password.route';
-export { ChangePassword } from './password/password.model';
-export { logoutRoute } from './logout/logout.route';
-export { LogoutComponent } from './logout/logout.component';
-export { PasswordSettingsComponent } from './password-settings/password-settings.component';
-export { IdpCallbackComponent } from './idp-callback/idp-callback.component';
-export { LoginErrorComponent } from './login-error/login-error.component';
-export { PasswordResetFinishComponent } from './password-reset/finish/password-reset-finish.component';
-export { PasswordResetFinish } from './password-reset/finish/password-reset-finish.service';
-export { passwordResetFinishRoute } from './password-reset/finish/password-reset-finish.route';
-export { PasswordResetInitComponent } from './password-reset/init/password-reset-init.component';
-export { PasswordResetInit } from './password-reset/init/password-reset-init.service';
-export { passwordResetInitRoute } from './password-reset/init/password-reset-init.route';
-export { passwordSetupRoute } from './password-reset/setup/password-setup.route';
-export { SignUpComponent } from './sign-up/sign-up.component';
-export { registerRoute } from './sign-up/sign-up.route';
-export { SettingsComponent } from './settings/settings.component';
-export { settingsRoute } from './settings/settings.route';
-export { idpCallbackRoute } from './idp-callback/idp-callback.route';
-export { loginErrorRoutes } from './login-error/login-error.route';
-export { IdleLogoutService } from './logout/idle-logout.service';
+export * from './activate';
+export * from './help';
+export * from './password';
+export * from './password-settings';
+export * from './logout';
+export * from './idp-callback';
+export * from './login-error';
+export * from './password-reset';
+export * from './sign-up';
+export * from './settings';
 export { GateAccountModule } from './account.module';
-export { accountState } from './account.route';
+export * from './route';
