+++ packages/account/login-error/index.ts
@@ -1,0 +1,2 @@
+export * from './login-error.component';
+export * from './login-error.route';
