+++ packages/account/logout/index.ts
@@ -1,0 +1,3 @@
+export * from './idle-logout.service';
+export * from './logout.component';
+export * from './logout.route';
