+++ packages/account/password-reset/finish/index.ts
@@ -1,0 +1,3 @@
+export * from './password-reset-finish.component';
+export * from './password-reset-finish.route';
+export * from './password-reset-finish.service';
