+++ packages/account/password-reset/init/index.ts
@@ -1,0 +1,3 @@
+export * from './password-reset-init.service';
+export * from './password-reset-init.component';
+export * from './password-reset-init.route';
