+++ packages/account/password-settings/index.ts
@@ -1,0 +1,2 @@
+export * from './password-settings.route';
+export * from './password-settings.component';
