+++ packages/account/password/index.ts
@@ -1,0 +1,4 @@
+export * from './password.component';
+export * from './password.model';
+export * from './password.module';
+export * from './password.service';
