+++ packages/account/route/account.route.ts
@@ -3,11 +3,12 @@ import { Routes } from '@angular/router';
 import { activateRoute } from '../activate/activate.route';
 import { helpRoute } from '../help/healp.route';
 import { idpCallbackRoute } from '../idp-callback/idp-callback.route';
+import { loginErrorRoutes } from '../login-error/login-error.route';
 import { logoutRoute } from '../logout/logout.route';
 import { passwordResetFinishRoute } from '../password-reset/finish/password-reset-finish.route';
 import { passwordResetInitRoute } from '../password-reset/init/password-reset-init.route';
 import { passwordSetupRoute } from '../password-reset/setup/password-setup.route';
-import { passwordRoute } from '../password/password.route';
+import { passwordRoute } from '../password-settings/password-settings.route';
 import { settingsRoute } from '../settings/settings.route';
 import { registerRoute } from '../sign-up/sign-up.route';
 
@@ -24,6 +25,6 @@ export const accountState: Routes = [{
         settingsRoute,
         logoutRoute,
         idpCallbackRoute,
-        // ...loginErrorRoutes,
+        ...loginErrorRoutes,
     ],
 }];
