+++ packages/account/route/index.ts
@@ -1,0 +1,1 @@
+export * from './account.route';
