+++ packages/account/settings/index.ts
@@ -1,0 +1,2 @@
+export * from './settings.route';
+export * from './settings.component';
