+++ packages/account/sign-up/index.ts
@@ -1,0 +1,2 @@
+export * from './sign-up.component';
+export * from './sign-up.route';
